requirements.txt
openai==1.65.2
PyQt6==6.8.1
python-dotenv==1.0.1
setuptools==58.0.4
termcolor==2.5.0
